refactor node agent
